+++ src/loader/utils.cpp
@@ -5,6 +5,7 @@
 #include <unordered_map>
 
 #include "src/common/include/configs.h"
+#include "src/storage/include/data_structure/lists/lists.h"
 
 namespace graphflow {
 namespace loader {
