+++ src/processor/include/physical_plan/operator/scan_attribute/scan_attribute.h
@@ -3,7 +3,7 @@
 #include "src/common/include/vector/node_vector.h"
 #include "src/processor/include/physical_plan/operator/physical_operator.h"
 #include "src/processor/include/task_system/morsel.h"
-#include "src/storage/include/structures/column.h"
+#include "src/storage/include/data_structure/column.h"
 
 using namespace graphflow::storage;
 
@@ -22,7 +22,7 @@ protected:
     shared_ptr<DataChunk> inDataChunk;
     shared_ptr<NodeIDVector> inNodeIDVector;
     shared_ptr<ValueVector> outValueVector;
-    unique_ptr<ColumnOrListsHandle> handle;
+    unique_ptr<DataStructureHandle> handle;
 };
 
 } // namespace processor
