+++ src/processor/include/physical_plan/operator/scan_attribute/scan_structured_column.h
@@ -3,7 +3,7 @@
 #include "src/common/include/vector/node_vector.h"
 #include "src/processor/include/physical_plan/operator/scan_attribute/scan_attribute.h"
 #include "src/processor/include/task_system/morsel.h"
-#include "src/storage/include/structures/column.h"
+#include "src/storage/include/data_structure/column.h"
 
 using namespace graphflow::storage;
