+++ src/processor/include/physical_plan/operator/tuple/data_chunks.h
@@ -2,7 +2,7 @@
 
 #include "src/common/include/data_chunk/data_chunk.h"
 #include "src/common/include/vector/node_vector.h"
-#include "src/storage/include/structures/common.h"
+#include "src/storage/include/data_structure/lists/list_sync_state.h"
 
 using namespace graphflow::common;
 using namespace graphflow::storage;
