+++ src/processor/physical_plan/operator/read_list/read_list.cpp
@@ -10,7 +10,7 @@ ReadList::ReadList(const uint64_t& dataChunkPos, const uint64_t& valueVectorPos,
     dataChunks = this->prevOperator->getDataChunks();
     inDataChunk = dataChunks->getDataChunk(dataChunkPos);
     inNodeIDVector = static_pointer_cast<NodeIDVector>(inDataChunk->getValueVector(valueVectorPos));
-    handle = make_unique<ColumnOrListsHandle>();
+    handle = make_unique<DataStructureHandle>();
 }
 
 void ReadList::readValuesFromList() {
