+++ src/processor/physical_plan/operator/scan_attribute/scan_unstructured_property.cpp
@@ -12,7 +12,7 @@ ScanUnstructuredProperty::ScanUnstructuredProperty(uint64_t dataChunkPos, uint64
     dataChunks = this->prevOperator->getDataChunks();
     inDataChunk = dataChunks->getDataChunk(dataChunkPos);
     inNodeIDVector = static_pointer_cast<NodeIDVector>(inDataChunk->getValueVector(valueVectorPos));
-    handle = make_unique<ColumnOrListsHandle>();
+    handle = make_unique<DataStructureHandle>();
     outValueVector = make_shared<ValueVector>(lists->getDataType());
     inDataChunk->append(outValueVector);
 }
