+++ src/storage/data_structure/data_structure.cpp
@@ -1,48 +1,23 @@
-#include "src/storage/include/structures/common.h"
+#include "src/storage/include/data_structure/data_structure.h"
 
 namespace graphflow {
 namespace storage {
 
-bool ListSyncState::hasNewRangeToRead() {
-    if (!hasValidRangeToRead()) {
-        return false;
-    }
-    if (startIdx + size == numElements) {
-        reset();
-        return false;
-    }
-    return true;
-}
-
-void ListSyncState::reset() {
-    startIdx = -1u;
-    size = -1u;
-    numElements = -1u;
-}
-
-bool ColumnOrListsHandle::hasMoreToRead() {
-    if (isAdjListsHandle) {
-        return listSyncState->hasNewRangeToRead();
-    } else {
-        return listSyncState->hasValidRangeToRead();
-    }
-}
-
-BaseColumnOrLists::BaseColumnOrLists(const string& fname, const DataType& dataType,
+DataStructure::DataStructure(const string& fname, const DataType& dataType,
     const size_t& elementSize, BufferManager& bufferManager)
     : logger{spdlog::get("storage")}, dataType{dataType}, elementSize{elementSize},
       numElementsPerPage{(uint32_t)(PAGE_SIZE / elementSize)}, fileHandle{fname},
       bufferManager(bufferManager){};
 
-void BaseColumnOrLists::reclaim(const unique_ptr<ColumnOrListsHandle>& handle) {
+void DataStructure::reclaim(const unique_ptr<DataStructureHandle>& handle) {
     if (handle->hasPageIdx()) {
         bufferManager.unpin(fileHandle, handle->getPageIdx());
         handle->resetPageIdx();
     }
 }
 
-void BaseColumnOrLists::readBySettingFrame(const shared_ptr<ValueVector>& valueVector,
-    const unique_ptr<ColumnOrListsHandle>& handle, PageCursor& pageCursor) {
+void DataStructure::readBySettingFrame(const shared_ptr<ValueVector>& valueVector,
+    const unique_ptr<DataStructureHandle>& handle, PageCursor& pageCursor) {
     const uint8_t* frame;
     if (handle->getPageIdx() != pageCursor.idx) {
         reclaim(handle);
@@ -54,8 +29,8 @@ void BaseColumnOrLists::readBySettingFrame(const shared_ptr<ValueVector>& valueV
     valueVector->values = ((uint8_t*)frame + pageCursor.offset);
 }
 
-void BaseColumnOrLists::readBySequentialCopy(const shared_ptr<ValueVector>& valueVector,
-    const unique_ptr<ColumnOrListsHandle>& handle, uint64_t sizeLeftToCopy, PageCursor& pageCursor,
+void DataStructure::readBySequentialCopy(const shared_ptr<ValueVector>& valueVector,
+    const unique_ptr<DataStructureHandle>& handle, uint64_t sizeLeftToCopy, PageCursor& pageCursor,
     unique_ptr<LogicalToPhysicalPageIdxMapper> mapper) {
     reclaim(handle);
     valueVector->reset();
