+++ src/storage/data_structure/lists/list_sync_state.cpp
@@ -1,0 +1,24 @@
+#include "src/storage/include/data_structure/lists/list_sync_state.h"
+
+namespace graphflow {
+namespace storage {
+
+bool ListSyncState::hasNewRangeToRead() {
+    if (!hasValidRangeToRead()) {
+        return false;
+    }
+    if (startIdx + size == numElements) {
+        reset();
+        return false;
+    }
+    return true;
+}
+
+void ListSyncState::reset() {
+    startIdx = -1u;
+    size = -1u;
+    numElements = -1u;
+}
+
+} // namespace storage
+} // namespace graphflow
