+++ src/storage/include/data_structure/lists/lists_metadata.h
@@ -6,7 +6,7 @@
 #include "spdlog/spdlog.h"
 
 #include "src/common/include/types.h"
-#include "src/storage/include/structures/common.h"
+#include "src/storage/include/data_structure/utils.h"
 
 using namespace std;
 using namespace graphflow::common;
@@ -135,56 +135,6 @@ private:
     // Total number of pages that is used to store the Lists data. This is the sum of number of
     // pages used by all chunks and large lists.
     uint32_t numPages;
-};
-
-// ListHeaders holds the headers of all lists in a single Lists structue.
-class ListHeaders {
-    friend class graphflow::loader::ListsLoaderHelper;
-    friend class graphflow::loader::AdjAndPropertyListsBuilder;
-    friend class graphflow::loader::NodesLoader;
-
-public:
-    ListHeaders() : logger{spdlog::get("storage")} {};
-    ListHeaders(string path);
-
-    uint32_t getHeader(node_offset_t offset) { return headers[offset]; };
-
-    static inline bool isALargeList(const uint32_t& header) { return header & 0x80000000; };
-
-    // For small lists.
-    static inline uint32_t getSmallListLen(const uint32_t& header) { return header & 0x7ff; };
-    static inline uint32_t getSmallListCSROffset(const uint32_t& header) {
-        return header >> 11 & 0xfffff;
-    };
-
-    // For large lists.
-    static inline uint32_t getLargeListIdx(const uint32_t& header) { return header & 0x7fffffff; };
-
-private:
-    void init(uint32_t size);
-
-    void saveToDisk(const string& fname);
-    void readFromDisk(const string& fname);
-
-private:
-    shared_ptr<spdlog::logger> logger;
-    // A header of a list is a uint32_t value the describes the following information about the
-    // list: 1) type: small or large, 2) location of list in pages.
-
-    // If the list is small, the layout of header is:
-    //      1. MSB (31st bit) is reset
-    //      2. 30-22 bits denote the index in the chunkPagesMaps[chunkId] where disk page ID is
-    //      located, where chunkId = idx of the list / LISTS_CHUNK_SIZE.
-    //      3. 21-11 bits denote the offset in the disk page.
-    //      4. 10-0  bits denote the number of elements in the list.
-
-    // If list is a large one (during initial data ingest, the criterion for being large is
-    // whether or not a list fits in a single page), the layout of header is:
-    //      1. MSB (31st bit) is set
-    //      2. 30-0  bits denote the idx in the largeListsPagesMaps where the length and the disk
-    //      page IDs are located.
-    unique_ptr<uint32_t[]> headers;
-    uint32_t size;
 };
 
 void saveListOfIntsToFile(const string& path, unique_ptr<uint32_t[]>& data, uint32_t listSize);
