+++ src/storage/include/store/nodes_store.h
@@ -5,8 +5,8 @@
 
 #include "src/common/include/types.h"
 #include "src/storage/include/catalog.h"
-#include "src/storage/include/structures/column.h"
-#include "src/storage/include/structures/lists.h"
+#include "src/storage/include/data_structure/column.h"
+#include "src/storage/include/data_structure/lists/lists.h"
 
 namespace graphflow {
 namespace storage {
