+++ src/storage/include/store/rels_store.h
@@ -2,8 +2,8 @@
 
 #include "src/common/include/types.h"
 #include "src/storage/include/catalog.h"
-#include "src/storage/include/structures/column.h"
-#include "src/storage/include/structures/lists.h"
+#include "src/storage/include/data_structure/column.h"
+#include "src/storage/include/data_structure/lists/lists.h"
 
 using namespace graphflow::common;
 using namespace std;
